染色体の値が取れる範囲をめっちゃ制限した。

+++ 手書き数字識別学習/ver 1/MINST.hpp
@@ -13,7 +13,7 @@ class MINST
         void LoadAns(char*, int);
 
         int ans;
-        std::vector<long double> data;
+        std::vector<double> data;
 };
 
 #endif
