+++ 手書き数字識別学習/ver 1/matrix.hpp
@@ -3,7 +3,7 @@
 
 #include <bits/stdc++.h>
 
-using Matrix = std::vector<std::vector<long double>>;
+using Matrix = std::vector<std::vector<double>>;
 class matrix
 {
     public:
